fix deamon and crc32

+++ deamons/sources/main/deamon_magic.h
@@ -15,6 +15,6 @@
 #include "deamonize.h"
 
 void start_server();
-size_t calc_filesize(char pathname[static 1]);
+signed long calc_filesize(char pathname[static 1]);
 
 #endif
